chore: Use snippet
